Refactor interfaces to soap package
