--- src/backend/libs/efiling-commons/src/main/java/ca/bc/gov/open/jag/efilingcommons/service/EfilingLookupService.java
@@ -1,13 +1,0 @@
-package ca.bc.gov.open.jag.efilingcommons.service;
-
-
-import ca.bc.gov.open.jag.efilingcommons.model.DocumentDetails;
-import ca.bc.gov.open.jag.efilingcommons.model.ServiceFees;
-
-import javax.xml.datatype.DatatypeConfigurationException;
-
-public interface EfilingLookupService {
-
-    ServiceFees getServiceFee(String serviceId);
-
-}
