+++ src/backend/libs/efiling-commons/src/main/java/ca/bc/gov/open/jag/efilingcommons/soap/service/EfilingLookupService.java
@@ -1,0 +1,10 @@
+package ca.bc.gov.open.jag.efilingcommons.soap.service;
+
+
+import ca.bc.gov.open.jag.efilingcommons.model.ServiceFees;
+
+public interface EfilingLookupService {
+
+    ServiceFees getServiceFee(String serviceId);
+
+}
